:recycle: Refactor FlowNodeParsers into smaller components
