+++ src/model/bpmn_model_parser.ts
@@ -1,13 +1,12 @@
+import * as xml2js from 'xml2js';
+
 import {
   IModelParser,
   IParsedObjectModel,
   Model,
 } from '@process-engine/process_model.contracts';
 
-import * as Parser from './parser';
-
-import * as BluebirdPromise from 'bluebird';
-import * as xml2js from 'xml2js';
+import {Parsers} from './parser';
 
 export class BpmnModelParser implements IModelParser {
 
@@ -22,9 +21,8 @@ export class BpmnModelParser implements IModelParser {
   };
 
   public async initialize(): Promise<void> {
-
     this._xmlParser = new xml2js.Parser(this.xmlParserOptions);
-    this._xmlParserFunc = BluebirdPromise.promisify(this._xmlParser.parseString, {
+    this._xmlParserFunc = Promise.promisify(this._xmlParser.parseString, {
       context: this._xmlParser,
     });
   }
@@ -32,7 +30,7 @@ export class BpmnModelParser implements IModelParser {
   public async parseXmlToObjectModel(xml: string): Promise<Model.Definitions> {
 
     const parsedObjectModel: IParsedObjectModel = await this._parseObjectModel(xml);
-    const definitions: Model.Definitions = Parser.parseDefinitions(parsedObjectModel);
+    const definitions: Model.Definitions = Parsers.DefinitionParser.parseDefinitions(parsedObjectModel);
 
     return definitions;
   }
