+++ src/model/index.ts
@@ -1,3 +1,3 @@
-export * from './parser';
+export * from './parser/index';
 export * from './bpmn_model_parser';
 export * from './type_factory';
