+++ src/model/parser/definitions_parser.ts
@@ -4,7 +4,7 @@ import {
   Model,
 } from '@process-engine/process_model.contracts';
 
-import * as Parser from './index';
+import {Parsers} from './index';
 
 export function parseDefinitions(parsedObjectModel: IParsedObjectModel): Model.Definitions {
 
@@ -17,8 +17,8 @@ function convertToInternalObjectModel(parsedXml: any): Model.Definitions {
 
   const definitions: Model.Definitions = createDefinitionBaseObject(parsedXml);
 
-  definitions.collaboration = Parser.parseCollaboration(parsedXml);
-  definitions.processes = Parser.parseProcesses(parsedXml);
+  definitions.collaboration = Parsers.CollaborationParser.parseCollaboration(parsedXml);
+  definitions.processes = Parsers.ProcessParser.parseProcesses(parsedXml);
 
   return definitions;
 }
