+++ src/model/parser/flow_node_parser.ts
@@ -1,9 +1,6 @@
 import {Model} from '@process-engine/process_model.contracts';
-import * as Parser from './index';
+import {FlowNodeParsers} from './flow_node_parsers/index';
 
-// TODO: The following elements are not supported yet:
-// - Text annotations
-// - Associations
 export function parseProcessFlowNodes(
   processData: any,
   errors: Array<Model.GlobalElements.Error>,
@@ -12,11 +9,12 @@ export function parseProcessFlowNodes(
 
   let nodes: Array<Model.Base.FlowNode> = [];
 
-  const events: Array<Model.Events.Event> = Parser.parseEventsFromProcessData(processData, errors, eventDefinitions);
-  const gateways: Array<Model.Gateways.Gateway> = Parser.parseGatewaysFromProcessData(processData);
-  const tasks: Array<Model.Activities.Activity> = Parser.parseActivitiesFromProcessData(processData, errors, eventDefinitions);
+  const events: Array<Model.Events.Event> = FlowNodeParsers.EventParser.parseEventsFromProcessData(processData, errors, eventDefinitions);
+  const gateways: Array<Model.Gateways.Gateway> = FlowNodeParsers.GatewayParser.parseGatewaysFromProcessData(processData);
+  const activities: Array<Model.Activities.Activity> =
+    FlowNodeParsers.ActivityParser.parseActivitiesFromProcessData(processData, errors, eventDefinitions);
 
-  nodes = nodes.concat(gateways, tasks, events);
+  nodes = nodes.concat(gateways, activities, events);
 
   return nodes;
 }
