+++ src/model/parser/flow_node_parsers/activity_parsers/activity_factory.ts
@@ -1,0 +1,19 @@
+import {BpmnTags, Model} from '@process-engine/process_model.contracts';
+
+import {
+  getModelPropertyAsArray,
+  setCommonObjectPropertiesFromData,
+} from '../../../type_factory';
+
+export function createActivityInstance<TActivity extends Model.Activities.Activity>(data: any, type: Model.Base.IConstructor<TActivity>): TActivity {
+
+  let instance: TActivity = new type();
+  instance = <TActivity> setCommonObjectPropertiesFromData(data, instance);
+
+  instance.incoming = getModelPropertyAsArray(data, BpmnTags.FlowElementProperty.SequenceFlowIncoming) || [];
+  instance.outgoing = getModelPropertyAsArray(data, BpmnTags.FlowElementProperty.SequenceFlowOutgoing) || [];
+
+  instance.name = data.name;
+
+  return instance;
+}
